refactor: saved search alerts routes
