+++ src/v2/Apps/SavedSearchAlerts/Components/SavedSearchAlertHeader.tsx
@@ -1,0 +1,9 @@
+import { Text } from "@artsy/palette"
+
+export const SavedSearchAlertHeader = () => {
+  return (
+    <Text variant="xl" my={4}>
+      Your Alerts
+    </Text>
+  )
+}
