--- src/v2/Apps/SavedSearchAlerts/Routes/Overview/types.ts
@@ -1,12 +1,0 @@
-export interface EditAlertEntity {
-  id: string
-  name: string
-  artistId: string
-}
-
-export interface EditFormProps {
-  editAlertEntity: EditAlertEntity
-  onCloseClick: () => void
-  onDeleteClick: () => void
-  onCompleted: () => void
-}
