+++ src/v2/Apps/SavedSearchAlerts/types.ts
@@ -1,0 +1,12 @@
+export interface EditAlertEntity {
+  id: string
+  name: string
+  artistId: string
+}
+
+export interface EditFormProps {
+  editAlertEntity: EditAlertEntity
+  onCloseClick: () => void
+  onDeleteClick: () => void
+  onCompleted: () => void
+}
